refactor package httputil to http
