+++ pkg/balance/metricsreader/backend_reader_test.go
@@ -21,7 +21,7 @@ import (
 	"github.com/pingcap/tiproxy/lib/util/logger"
 	"github.com/pingcap/tiproxy/lib/util/waitgroup"
 	"github.com/pingcap/tiproxy/pkg/manager/infosync"
-	"github.com/pingcap/tiproxy/pkg/util/httputil"
+	httputil "github.com/pingcap/tiproxy/pkg/util/http"
 	dto "github.com/prometheus/client_model/go"
 	"github.com/prometheus/common/model"
 	"github.com/stretchr/testify/require"
